Update `diol` dependency
While just poking around, I stumbled across this repository and tried to
run the benchmarks. It failed due to a `fontconfig` issue which I
tracked down to the `diol` dependency. In more recent times, `diol` has
gained a `plot` feature that I was able to turn off to get things to
run. In doing so, I upgraded `diol` to its latest version and thought to
upstream it--this PR. Feel free to take it or leave it or change it as
you see fit!